requirements.txt
annotated-types
anyio
beautifulsoup4
bs4
certifi
charset-normalizer
click
colorama
distro
fastapi
fuzzywuzzy
h11
httpcore
httpx
idna
jiter
Levenshtein
openai
pydantic
pydantic_core
python-Levenshtein
RapidFuzz
requests
sniffio
soupsieve
starlette
tqdm
typing_extensions
urllib3
uvicorn
pillow
python-dateutil
python-dotenv
numpy
duckdb
markdown
faker
